Create HTML Version
Started teaching myself some HTML, this definitely seems like a preferable option.  I'm having trouble booting it for some reason or other, possibly to do with my virus protection or something.  No-less, it should have most of the functionalities we need save for a history.  This is also lacking in the style department if someone wants to dabble with CSS to pretty it up.